src/routes/transactions.ts
import { FastifyInstance } from 'fastify'
import { z } from 'zod'
import { knex } from '../database'
import { randomUUID } from 'node:crypto'
import { checkSessionIdExists } from '../middlewares/check-session-id-exists'

export async function transactionsRoutes(app: FastifyInstance) {
  app.get('/', { preHandler: [checkSessionIdExists] }, async (request) => {
    const { sessionId } = request.cookies

    const transactions = knex('transactions')
      .where({
        session_id: sessionId,
      })
      .select('*')

    return transactions
  })

  app.get('/:id', { preHandler: [checkSessionIdExists] }, async (request) => {
    const { sessionId } = request.cookies

    const requestParamSchema = z.object({
      id: z.string().uuid(),
    })

    const { id } = requestParamSchema.parse(request.params)

    const transaction = knex('transactions')
      .where({
        id,
        session_id: sessionId,
      })
      .first()

    return transaction
  })

  app.get(
    '/summary',
    { preHandler: [checkSessionIdExists] },
    async (request) => {
      const { sessionId } = request.cookies

      const summary = knex('transactions')
        .where({
          session_id: sessionId,
        })
        .sum('amount', { as: 'summary' })
        .first()

      return summary
    },
  )

  app.post('/', async (request, response) => {
    const requestBodySchema = z.object({
      title: z.string(),
      amount: z.number(),
      type: z.enum(['credit', 'debit']),
    })

    const { title, amount, type } = requestBodySchema.parse(request.body)

    let sessionId = request.cookies.sessionId

    if (!sessionId) {
      sessionId = randomUUID()

      response.cookie('sessionId', sessionId, {
        path: '/',
        maxAge: 1000 * 60 * 60 * 24 * 7, // 7 days
      })
    }

    await knex('transactions').insert({
      id: randomUUID(),
      title,
      amount: type === 'credit' ? amount : amount * -1,
      session_id: sessionId,
    })

    response.status(201).send()
  })
}
